gather defaults in one place
