+++ packages/cli/src/schemas/v1.1/schema.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
-import { baseColorNames, colorNames } from '../../colors/colorMetadata.ts';
+import { colorNames } from '../../colors/colorMetadata.ts';
 import { convertToHex } from '../../colors/index.ts';
-import { defaultBorderRadius, defaultFontFamily } from '../defaults.ts';
+import { defaultBorderRadius, defaultFontFamily, severityColorNames } from '../defaults.ts';
 
 const hexPatterns = [
   // Hex colors: #000, #0000, #000000, #00000000
@@ -35,7 +35,10 @@ const semanticColorOverrideSchema = z
   .describe('An object with color names as keys');
 
 const severityColorOverrideSchema = z
-  .partialRecord(z.enum(baseColorNames), colorSchema.describe('A hex color, which is used for creating a color scale'))
+  .partialRecord(
+    z.enum(severityColorNames),
+    colorSchema.describe('A hex color, which is used for creating a color scale'),
+  )
   .optional()
   .describe('An object with severity color names as keys');
