+++ packages/cli/src/schemas/v1/schema.ts
@@ -1,7 +1,8 @@
 import { z } from 'zod';
-import { baseColorNames, colorNames } from '../../colors/colorMetadata.ts';
+import { colorNames } from '../../colors/colorMetadata.ts';
 import { convertToHex } from '../../colors/index.ts';
 import { RESERVED_COLORS } from '../../colors/theme.ts';
+import { severityColorNames } from '../defaults.ts';
 
 const hexPatterns = [
   // Hex colors: #000, #0000, #000000, #00000000
@@ -53,7 +54,10 @@ const semanticColorOverrideSchema = z
   .describe('An object with color names as keys');
 
 const severityColorOverrideSchema = z
-  .partialRecord(z.enum(baseColorNames), colorSchema.describe('A hex color, which is used for creating a color scale'))
+  .partialRecord(
+    z.enum(severityColorNames),
+    colorSchema.describe('A hex color, which is used for creating a color scale'),
+  )
   .optional()
   .describe('An object with severity color names as keys');
